Moved query out of event
