+++ quic/s2n-quic-core/src/lib.rs
@@ -22,6 +22,7 @@ pub mod io;
 pub mod number;
 pub mod packet;
 pub mod path;
+pub mod query;
 pub mod random;
 pub mod recovery;
 pub mod slice;
