+++ quic/s2n-quic-transport/src/connection/api.rs
@@ -16,8 +16,8 @@ use core::{
 use s2n_quic_core::{
     application,
     application::ServerName,
-    event::query::{Query, QueryMut},
     inet::SocketAddress,
+    query::{Query, QueryMut},
     stream::StreamType,
 };
