+++ quic/s2n-quic-transport/src/connection/connection_container/tests.rs
@@ -30,6 +30,7 @@ use s2n_quic_core::{
         zero_rtt::ProtectedZeroRtt,
     },
     path::MaxMtu,
+    query,
     time::{Timer, Timestamp},
 };
 use std::sync::Mutex;
@@ -305,15 +306,15 @@ impl connection::Trait for TestConnection {
         None
     }
 
-    fn query_event_context(&self, _query: &mut dyn event::query::Query) {
+    fn query_event_context(&self, _query: &mut dyn query::Query) {
         todo!()
     }
 
-    fn query_event_context_mut(&mut self, _query: &mut dyn event::query::QueryMut) {
+    fn query_event_context_mut(&mut self, _query: &mut dyn query::QueryMut) {
         todo!()
     }
 
-    fn datagram_mut(&mut self, _query: &mut dyn event::query::QueryMut) {
+    fn datagram_mut(&mut self, _query: &mut dyn query::QueryMut) {
         todo!()
     }
